USHIFT-1436: only download images if the compose completes successfully

+++ test/bin/download_images.sh
@@ -12,12 +12,22 @@ mkdir -p "${VM_DISK_DIR}"
 
 download_image() {
     local buildid="${1}"
+    local status
+
+    status=$(sudo composer-cli compose status | grep "${buildid}" | awk '{print $2}')
+    if [ "${status}" != "FINISHED" ]; then
+        sudo composer-cli compose info "${buildid}"
+        return 1
+    fi
+
     # shellcheck disable=SC2086  # pass glob args without quotes
     rm -f ${buildid}-*.tar
     sudo composer-cli compose metadata "${buildid}"
     sudo composer-cli compose image "${buildid}"
     # shellcheck disable=SC2086  # pass glob args without quotes
     sudo chown "$(whoami)." ${buildid}-*
+
+    return 0
 }
 
 download_dir="${IMAGEDIR}/builds"
@@ -28,7 +38,10 @@ cd "${download_dir}"
 for blueprint_build in *.image-installer; do
     blueprint=$(basename "${blueprint_build}" .image-installer)
     buildid=$(cat "${blueprint_build}")
-    download_image "${buildid}"
+    if ! download_image "${buildid}"; then
+        echo "Did not download image"
+        continue
+    fi
     iso_file="${buildid}-installer.iso"
     echo "Moving ${iso_file} to ${VM_DISK_DIR}/${blueprint}.iso"
     mv -f "${iso_file}" "${VM_DISK_DIR}/${blueprint}.iso"
@@ -44,7 +57,10 @@ cd "${download_dir}"
 for blueprint_build in *.edge-commit; do
     blueprint=$(basename "${blueprint_build}" .edge-commit)
     buildid=$(cat "${blueprint_build}")
-    download_image "${buildid}"
+    if ! download_image "${buildid}"; then
+        echo "Did not download image"
+        continue
+    fi
     commit_file="${buildid}-commit.tar"
     echo "Unpacking ${commit_file} ${blueprint}"
     tar -C "${IMAGEDIR}" -xf "${commit_file}"
